Decouple `CheckLintNameResult` from `rustc_lint`
